Continuing improvements

+++ src/alpha_check.rs
@@ -85,7 +85,7 @@ fn has_non_constant_cap_alpha_f32_impl<const ALPHA_CHANNEL_INDEX: usize, const C
     if store.is_empty() {
         return false;
     }
-    let first = store[0].to_bits() as u32;
+    let first = store[0].to_bits();
     let mut row_sums: u64 = 0u64;
     for row in store.chunks_exact(width * CHANNELS) {
         for color in row.chunks_exact(CHANNELS) {
